Add authorization to endpoints

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -55,11 +56,12 @@ public class AuctionsController : ControllerBase
     }
 
     [HttpPost]
+    [Authorize]
     public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
     {
         var auction = _mapper.Map<Auction>(auctionDto);
-        // TODO: Add current user as seller
-        auction.Seller = "Test";
+        
+        auction.Seller = User.Identity.Name;
         auction.AuctionEnd = auction.AuctionEnd.ToUniversalTime();
         
         var newAuction = _mapper.Map<AuctionDto>(auction);
@@ -79,6 +81,7 @@ public class AuctionsController : ControllerBase
             newAuction);
     }
 
+    [Authorize]
     [HttpPut("{id}")]
     public async Task<ActionResult> UpdateAuction(
         Guid id, 
@@ -89,9 +92,10 @@ public class AuctionsController : ControllerBase
 
         if (auction == null)
             return NotFound();
-        
-        
-        //TODO: Check seller == username
+
+
+        if (auction.Seller != User.Identity.Name)
+            return Forbid();
 
         auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
         auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
@@ -109,6 +113,7 @@ public class AuctionsController : ControllerBase
         return Ok();
     }
 
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<ActionResult> DeleteAuction(Guid id)
     {
@@ -116,8 +121,9 @@ public class AuctionsController : ControllerBase
 
         if (auction is null)
             return NotFound();
-        
-        // TODO: Check seller == username
+
+        if (auction.Seller != User.Identity.Name)
+            return Forbid();
 
         _context.Auctions.Remove(auction);
